Initial fixes from code review
