+++ packages/front-end/services/dates.ts
@@ -18,12 +18,6 @@ export function ago(date: string | Date): string {
 export function daysLeft(date: string | Date): number {
   return differenceInDays(getValidDate(date), new Date());
 }
-export function daysBetween(
-  date1: string | Date,
-  date2: string | Date
-): number {
-  return differenceInDays(getValidDate(date2), getValidDate(date1));
-}
 export function subtractMonths(date: string | Date, num: number): Date {
   return addMonths(getValidDate(date), -1 * num);
 }
